better position controller. physical issue: solonoid pulled too much current

+++ Node2/pwm.h
@@ -3,5 +3,5 @@
 
 void pwm_init();
 void pwm_set_duty_cycle(uint16_t duty_cycle);
-void pwm_set_position(uint8_t pos);
+void pwm_set_position(int pos);
 void motor_set_duty_cycle(uint16_t duty_cycle);
